Addition of tests and Task5.

+++ lab_01_03_03/main.c
@@ -1,22 +1,22 @@
 #include <stdio.h>
 #include <math.h>
 
-float count_moving(float speed, float acceleration, float time);
+double count_moving(double speed, double acceleration, double time);
 
 int main(void)
 {
-    float speed, acceleration, time;
-    float moving;
+    double speed, acceleration, time;
+    double moving;
 
-    scanf("%f%f%f", &speed, &acceleration, &time);
+    scanf("%lf%lf%lf", &speed, &acceleration, &time);
     moving = count_moving(speed, acceleration, time);
     printf("%f\n", moving);
 
     return 0;
 }
 
-float count_moving(float speed, float acceleration, float time)
+double count_moving(double speed, double acceleration, double time)
 {
-    return speed*time + acceleration*pow(time, 2)/2;
+    return speed * time + acceleration * pow(time, 2) / 2;
 }
